requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
idna==3.3
oauthlib==3.1.1
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
